max30105 sensor reading changes

+++ include/getSensorData.h
@@ -7,10 +7,10 @@
 uint16_t getGsrData();
 uint16_t getIrSensorValue();
 uint16_t getBpm();
-uint16_t getAvgBpm();
+int getAvgBpm();
 
 void getHeartRateData();
 void initMax();
-void sensor_to_container(uint8_t id, String name, uint16_t value_1, uint16_t value_2 = 0, uint16_t value_3 = 0);
+void sensor_to_container(uint8_t id, String name, uint16_t value_1, int value_2 = 0, uint16_t value_3 = 0);
 
 #endif
